feat: ItemRowView의 가격이 소수점 이하는 버리고 정수로 표시되도록 구현
- ItemDTO의 Double 타입 가격이 Entity로 바뀌면서 소수점 이하는 버리고 정수가 되도록 구현

+++ HongdangmuMarket/HongdangmuMarket/Models/Entities/Item.swift
@@ -12,7 +12,7 @@ struct Item {
     let id, vendorID: Int
     let name: String
     let thumbnail: String
-    let price, bargainPrice, discountedPrice: Double
+    let price, bargainPrice, discountedPrice: Int
     let stock: Int
     let createdAt, issuedAt: String
